% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust}
\alias{p_adjust}
\title{P-value Corrections}
\description{This function executes an specific p-value adjustment method.}
\arguments{
  \item{data (dataframe)} {dataframe containing at least a column of p-values to be adjusted}
  \item{pv_index (int)} {if the input is a dataframe, col refers to the column name of the dataframe that has the p-values.}
  \item{method (numeric)} {significance level for both methods.}
  \item{alpha (string)} {method used for the adjustment ("bh" or "bonf")}
           }
\usage{p_adjust(data, pv_index, method, alpha=0.05)}
\examples{
Adapted from https://matthew-brett.github.io/teaching/fdr.html
nSims <- 100 #number of simulated experiments
p <-numeric(nSims) #set up empty container for all simulated p-values

for(i in 1:nSims){ #for each simulated experiment
  x<-rnorm(n = 100, mean = 10, sd = 8) #produce 100 simulated participants
  #with mean=10 and SD=8
  y<-rnorm(n = 100, mean = 30, sd = 5) #produce 100 simulated participants
  #with mean=30 and SD=5
  z<-t.test(x,y) #perform the t-test
  p[i]<-z$p.value #get the p-value and store it
}
p_adjust(p ,1, "bonf",0.05)
}
