% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_methods.R
\name{p_methods}
\alias{p_methods}
\title{P-value Corrections}
\usage{
p_methods(data, pv_index, alpha = 0.05)
}
\arguments{
\item{data}{(dataframe) dataframe containing at least a column of p-values to be adjusted}

\item{pv_index}{(numeric) if the input is a dataframe, col refers to the column name of the dataframe that has the p-values}

\item{alpha}{(numeric) significance level for both methods}
}
\value{
data.frame
}
\description{
Dataframe: appends to input dataframe both adjusted p-values and significance
                 levels (Bonferroni and BH) in ascending raw p-value order.
     Includes following columns:
     - bonf_value (numeric): Bonferroni adjusted significance level (same for all)
     - bonf_significant (logical): True if significant p-value or False if not
     - bh_value (numeric): Benjamini-Hochberg (BH) critical value
     - BH_significant (logical): True if significant p-value or False if not
}
\details{
Requirements:  Dplyr

Requirements:  Dplyr
}
\examples{
# test function
p_method(data = p, pv_index = 1, alpha = 0.05)
}
