% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptoolkit-package.r
\docType{package}
\name{ptoolkit}
\alias{ptoolkit}
\alias{ptoolkit-package}
\title{ptoolkit.}
\description{
p_toolkit is a package designed to help adjust and visualize p-values when using multiple comparisons.
As computing power has become powerful enough to run hundreds or even thousands of statistical tests,
it is important to look at small p-values and try to understand whether the result is small simply by chance,
or whether it truly is significant.
}
