% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust}
\alias{p_adjust}
\title{P-value Adjustments by Methods}
\usage{
p_adjust(data, pv_index, method, alpha = 0.05)
}
\arguments{
\item{data}{(dataframe) dataframe containing at least a column of p-values to be adjusted}

\item{pv_index}{(numeric) if the input is a dataframe, col refers to the column name of the dataframe that has the p-values}

\item{method}{(string) method used for the adjustment ("bh" or "bonf")}

\item{alpha}{(numeric) significance level for both methods}
}
\description{
This function outputs a dataframe with the original p-values and their adjustments by the specified method.
}
\details{
Requirements:  Dplyr
}
\examples{
# test function
p_adjust(data = [0.03,0.05,0.08,0.2], pv_index = 1, method = "bonf", alpha = 0.05)

}
