% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust}
\alias{p_adjust}
\title{P-value Adjustments
This function executes an specific p-value adjustment method.
Args:}
\usage{
p_adjust(data, pv_index, method, alpha = 0.05)
}
\arguments{
\item{data}{(dataframe): dataframe containing at least a column of p-values to be adjusted}

\item{pv_index}{(int): if the input is a dataframe, col refers to the column name of the dataframe that has the p-values.}

\item{method}{(numeric): significance level for both methods.}

\item{alpha}{(string): method used for the adjustment ("bh" or "bonf")}
}
\description{
P-value Adjustments
This function executes an specific p-value adjustment method.
Args:
}
\examples{

# Trial




}
