% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plot.R
\name{p_plot}
\alias{p_plot}
\title{P-value Summary Plot}
\usage{
p_plot(data, pv_index, alpha = 0.05)
}
\arguments{
\item{data}{dataframe containing at least a column of p-values to be adjusted}

\item{pv_index}{(numeric) if the input is a dataframe, col refers to the column name of the dataframe that has the p-values}

\item{alpha}{(numeric) significance level for both methods}
}
\value{
ggplot dataframe output from the p_methods function
}
\description{
This function plots all the p-values in ascending order and compares them with two lines, one representing the BH cutoff point and another one the Bonferroni cutoff.
}
