% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_plot.R
\name{p_plot}
\alias{p_plot}
\title{This function plots all the p-values in ascending order and compares them with two lines, one representing
the BH cutoff point and another one the Bonferroni cutoff.}
\usage{
p_plot(data, pv_index, alpha = 0.05)
}
\arguments{
\item{data}{}
}
\value{
ggplot2
}
\description{
Args:
  - ad_object: the dataframe output from the p_methods function.
}
\details{
Returns:
  - plot: a ggplot with the p-values and both cut-off lines.
Title
}
